Major reorganization of code. Modules now include main to be runned as simple scripts
